aero_timestamp_asr_pypi/pyproject.toml
[tool.black]
line-length = 240

[build-system]
requires = ["setuptools>=42", "wheel", "setuptools_scm[tomli]>=6.3"]
build-backend = "setuptools.build_meta"

[project]
name = "aero_timestamp_asr"
version = "0.1.0"
authors = [
    { name = "Kaichen, Zhang", email = "kaichenzhang358@outlook.com" },
]
description = "Aero ASR timestamp package"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
license = { text = "MIT" }
dependencies = [
    "transformers==4.51.3",
    "hf_transfer",
    "dtw-python",
    "accelerate",
    "torch==2.5.1",
    "librosa",
    "soundfile",
]


[tool.setuptools.packages.find]
include = ["aero_timestamp_asr*"]
exclude = [
    "test"
]

[tool.wheel]
exclude = [
    "test"
]

[project.urls]
Homepage = "https://www.lmms-lab.com/posts/aero_audio/"
Repository = "https://github.com/EvolvingLMMs-Lab/Aero-1"